+++ ztf_viewer/catalogs/__init__.py
@@ -1,2 +1,2 @@
 from .unavailable_catalogs import unavailable_catalogs
-from .ztf_dr import find_ztf_oid, find_ztf_circle
+from .ztf_dr import find_ztf_circle, find_ztf_oid