+++ ztf_viewer/catalogs/conesearch/__init__.py
@@ -11,30 +11,29 @@ from .gcvs import GcvsQuery
 from .ogle import OgleQuery
 from .panstarrs import PanstarrsDr2StackedQuery
 from .sdss import SdssQuasarsQuery
-from .spicy import SpicyQuery
 from .simbad import SimbadQuery
+from .spicy import SpicyQuery
 from .tns import TnsQuery
 from .vsx import VsxQuery
 from .ztf_periodic import ZtfPeriodicQuery
 
-
-ALERCE_QUERY = AlerceQuery('Alerce')
-ANTARES_QUERY = AntaresQuery('Antares')
-ASTROCATS_QUERY = AstrocatsQuery('Astrocats')
-FINK_QUERY = FinkQuery('Fink')
-ATLAS_QUERY = AtlasQuery('ATLAS')
-GAIA2_DIS = Gaia2Dis('Gaia DR2 Distances')
-GAIA_DR3 = GaiaDr3Query('Gaia DR3')
-GAIA_EDR3_DIS = GaiaEdr3Dis('Gaia EDR3 Distances')
-GCVS_QUERY = GcvsQuery('GCVS')
-OGLE_QUERY = OgleQuery('OGLE')
-PANSTARRS_DR2_QUERY = PanstarrsDr2StackedQuery('Pan-STARRS DR2 Stacked')
-SDSS_QUASARS_QUERY = SdssQuasarsQuery('SDSS DR16 Quasars')
-SIMBAD_QUERY = SimbadQuery('Simbad')
-SPICY_QUERY = SpicyQuery('SPICY')
-TNS_QUERY = TnsQuery('Transient Name Server')
-VSX_QUERY = VsxQuery('VSX')
-ZTF_PERIODIC_QUERY = ZtfPeriodicQuery('ZTF Periodic')
+ALERCE_QUERY = AlerceQuery("Alerce")
+ANTARES_QUERY = AntaresQuery("Antares")
+ASTROCATS_QUERY = AstrocatsQuery("Astrocats")
+FINK_QUERY = FinkQuery("Fink")
+ATLAS_QUERY = AtlasQuery("ATLAS")
+GAIA2_DIS = Gaia2Dis("Gaia DR2 Distances")
+GAIA_DR3 = GaiaDr3Query("Gaia DR3")
+GAIA_EDR3_DIS = GaiaEdr3Dis("Gaia EDR3 Distances")
+GCVS_QUERY = GcvsQuery("GCVS")
+OGLE_QUERY = OgleQuery("OGLE")
+PANSTARRS_DR2_QUERY = PanstarrsDr2StackedQuery("Pan-STARRS DR2 Stacked")
+SDSS_QUASARS_QUERY = SdssQuasarsQuery("SDSS DR16 Quasars")
+SIMBAD_QUERY = SimbadQuery("Simbad")
+SPICY_QUERY = SpicyQuery("SPICY")
+TNS_QUERY = TnsQuery("Transient Name Server")
+VSX_QUERY = VsxQuery("VSX")
+ZTF_PERIODIC_QUERY = ZtfPeriodicQuery("ZTF Periodic")
 
 
 catalog_query_objects = _BaseCatalogQuery.get_objects