+++ ztf_viewer/date_with_frac.py
@@ -8,7 +8,7 @@ import requests
 
 from ztf_viewer.cache import cache
 from ztf_viewer.config import ZTF_FITS_PROXY_URL
-from ztf_viewer.util import hmjd_to_earth, qid_from_rcid, ccdid_from_rcid
+from ztf_viewer.util import ccdid_from_rcid, hmjd_to_earth, qid_from_rcid
 
 
 @dataclass
@@ -31,23 +31,23 @@ class DateWithFrac:
 
     @property
     def monthday(self):
-        return f'{self.month:02d}{self.day:02d}'
+        return f"{self.month:02d}{self.day:02d}"
 
     def frac_digits(self, digits):
         return round(self.fraction * 10**digits)
 
     @property
     def products_root(self):
-        return f'/products/sci/{self.year}/{self.monthday}/'
+        return f"/products/sci/{self.year}/{self.monthday}/"
 
     @property
     def products_path(self):
-        return f'{self.products_root}{self.frac_digits(6):06d}/'
+        return f"{self.products_root}{self.frac_digits(6):06d}/"
 
     def sciimg_path(self, *, fieldid, filter, rcid):
         ccdid = ccdid_from_rcid(rcid)
         qid = qid_from_rcid(rcid)
-        filename = f'ztf_{self.year}{self.monthday}{self.frac_digits(6):06d}_{fieldid:06d}_{filter}_c{ccdid:02d}_o_q{qid}_sciimg.fits'
+        filename = f"ztf_{self.year}{self.monthday}{self.frac_digits(6):06d}_{fieldid:06d}_{filter}_c{ccdid:02d}_o_q{qid}_sciimg.fits"
         return os.path.join(self.products_path, filename)
 
 
@@ -63,4 +63,4 @@ def correct_date(date_with_frac):
     fracs = _fracs(date_with_frac.products_root)
     digits = 6
     i = np.searchsorted(fracs, date_with_frac.frac_digits(digits))
-    date_with_frac.fraction = fracs[i - 1] / (10.0 ** digits)
+    date_with_frac.fraction = fracs[i - 1] / (10.0**digits)