+++ ztf_viewer/pages/favicon.py
@@ -3,6 +3,6 @@ from flask import send_file
 from ztf_viewer.app import app
 
 
-@app.server.route('/favicon.ico')
+@app.server.route("/favicon.ico")
 def favicon():
-    return send_file('static/img/logo.svg', mimetype='image/svg+xml')
+    return send_file("static/img/logo.svg", mimetype="image/svg+xml")