+++ ztf_viewer/pages/lc_csv.py
@@ -16,31 +16,31 @@ def get_csv(dr, oids):
             raise NotFound
         meta = find_ztf_oid.get_meta(oid, dr)
         oid_df = pd.DataFrame.from_records(lc)
-        oid_df['oid'] = oid
-        oid_df['filter'] = meta['filter']
+        oid_df["oid"] = oid
+        oid_df["filter"] = meta["filter"]
         dfs.append(oid_df)
-    df = pd.concat(dfs, axis='index')
-    df.sort_values(by='mjd', inplace=True)
-    df = df[['oid', 'filter', 'mjd', 'mag', 'magerr', 'clrcoeff']]
+    df = pd.concat(dfs, axis="index")
+    df.sort_values(by="mjd", inplace=True)
+    df = df[["oid", "filter", "mjd", "mag", "magerr", "clrcoeff"]]
 
     string_io = StringIO()
     df.to_csv(string_io, index=False)
     return string_io.getvalue()
 
 
-@app.server.route('/<dr>/csv/<int:oid>')
+@app.server.route("/<dr>/csv/<int:oid>")
 def response_csv(dr, oid):
     try:
-        other_oids = set(map(int, request.args.getlist('other_oid')))
+        other_oids = set(map(int, request.args.getlist("other_oid")))
     except ValueError:
-        return 'other_oid query parameter must be an integer', 400
+        return "other_oid query parameter must be an integer", 400
     oids = set.union({oid}, other_oids)
     try:
         csv = get_csv(dr, oids)
     except NotFound:
-        return '', 404
+        return "", 404
     return Response(
         csv,
-        mimetype='text/csv',
-        headers={'Content-disposition': f'attachment; filename={oid}.csv'},
+        mimetype="text/csv",
+        headers={"Content-disposition": f"attachment; filename={oid}.csv"},
     )