+++ ztf_viewer/pages/login.py
@@ -1,5 +1,5 @@
 import dash
-from dash import html, dcc, Input, Output, State
+from dash import Input, Output, State, dcc, html
 
 from ztf_viewer.akb import akb
 from ztf_viewer.app import app
@@ -10,31 +10,28 @@ def get_layout(pathname):
     return html.Div(
         [
             html.Br(),
-            'Token:',
+            "Token:",
             dcc.Input(
-                id='token',
-                type='text',
+                id="token",
+                type="text",
                 minLength=8,
                 maxLength=40,
                 n_submit=0,
             ),
-            html.Div('', id='login-status')
+            html.Div("", id="login-status"),
         ],
-        id='login',
+        id="login",
     )
 
 
-@app.callback(
-    Output('login-status', 'children'),
-    [Input('token', 'n_submit')],
-    [State('token', 'value')]
-)
+@app.callback(Output("login-status", "children"), [Input("token", "n_submit")], [State("token", "value")])
 def do_login(n_submit, token):
     try:
         username = akb.username(token)
     except UnAuthorized:
-        return 'Login failed: wrong token'
+        return "Login failed: wrong token"
     if token:
-        dash.callback_context.response.set_cookie('akb_token', token, secure=True, samesite='Strict',
-                                                  max_age=31 * 86400)
-    return ['You are authorised as ', html.B(username)]
+        dash.callback_context.response.set_cookie(
+            "akb_token", token, secure=True, samesite="Strict", max_age=31 * 86400
+        )
+    return ["You are authorised as ", html.B(username)]