+++ ztf_viewer/ttl_set.py
@@ -1,9 +1,9 @@
-import packaging.version
 import pickle
 from abc import ABC, abstractmethod
 from collections.abc import MutableSet
-from typing import Hashable, Iterator, TypeVar, Dict, Any, Generic
+from typing import Any, Dict, Generic, Hashable, Iterator, TypeVar
 
+import packaging.version
 from cachetools import TTLCache
 from redis import StrictRedis
 
@@ -38,7 +38,7 @@ class LocalTTLSet(_BaseTTLSet[_T_Local], Generic[_T_Local]):
     def __init__(self, maxsize: int, ttl: int):
         super().__init__(ttl)
         self.ttl_cache = TTLCache(maxsize=maxsize, ttl=ttl)
-        
+
     def add(self, value: _T_Local) -> None:
         self.ttl_cache[value] = None
 
@@ -63,17 +63,17 @@ _T_Redis = TypeVar("_T_Redis")
 
 
 class RedisTTLSet(_BaseTTLSet[_T_Redis], Generic[_T_Redis]):
-    def __init__(self, ttl: int, client: StrictRedis, prefix: str = 'RedisTTLSet'):
+    def __init__(self, ttl: int, client: StrictRedis, prefix: str = "RedisTTLSet"):
         super().__init__(ttl)
 
         self.client = client
 
         self.prefix = prefix.encode()
-        if b'*' in self.prefix:
+        if b"*" in self.prefix:
             raise ValueError('prefix must not contain "*"')
 
         self.redis_version = self.__redis_version(client)
-        if self.redis_version < packaging.version.parse('6.2.0'):
+        if self.redis_version < packaging.version.parse("6.2.0"):
             self.remove = self.__remove_pre_6_2_0
         else:
             self.remove = self.__remove_6_2_0
@@ -81,7 +81,7 @@ class RedisTTLSet(_BaseTTLSet[_T_Redis], Generic[_T_Redis]):
     @staticmethod
     def __redis_version(client: StrictRedis) -> packaging.version.Version:
         info = client.info()
-        version = packaging.version.parse(info['redis_version'])
+        version = packaging.version.parse(info["redis_version"])
         return version
 
     def _encode(self, value: _T_Redis) -> bytes:
@@ -102,16 +102,16 @@ class RedisTTLSet(_BaseTTLSet[_T_Redis], Generic[_T_Redis]):
     def __remove_pre_6_2_0(self, value: _T_Redis) -> None:
         key = self._encode(value)
         if self.client.exists(key) == 0:
-            raise KeyError(f'{value} not found')
+            raise KeyError(f"{value} not found")
         self.client.delete(key)
 
     def __remove_6_2_0(self, value: _T_Redis) -> None:
         key = self._encode(value)
         if self.client.getdel(key) is None:
-            raise KeyError(f'{value} not found')
+            raise KeyError(f"{value} not found")
 
     def clear(self) -> None:
-        for key in self.client.keys(self.prefix + b'*'):
+        for key in self.client.keys(self.prefix + b"*"):
             self.client.delete(key)
 
     def add(self, value: _T_Redis) -> None:
@@ -123,10 +123,10 @@ class RedisTTLSet(_BaseTTLSet[_T_Redis], Generic[_T_Redis]):
         return self.client.exists(key) > 0
 
     def __len__(self) -> int:
-        return len(self.client.keys(self.prefix + b'*'))
+        return len(self.client.keys(self.prefix + b"*"))
 
     def __iter__(self) -> Iterator[_T_Redis]:
-        for key in self.client.keys(self.prefix + b'*'):
+        for key in self.client.keys(self.prefix + b"*"):
             yield self._decode(key)